Colocate tests and start on view from index
