--- src/components/publishing/__test__/sections.test.tsx
@@ -1,14 +1,0 @@
-import * as React from "react"
-import * as renderer from "react-test-renderer"
-
-import "jest-styled-components"
-
-import Sections from "../sections/sections"
-import Articles from "./fixtures/articles"
-
-jest.mock("react-sizeme", () => jest.fn(c => d => d))
-
-it("renders properly", () => {
-  const sections = renderer.create(<Sections article={Articles[1]} />).toJSON()
-  expect(sections).toMatchSnapshot()
-})
