+++ src/components/publishing/sections/__test__/authors.test.tsx
@@ -3,8 +3,8 @@ import * as renderer from "react-test-renderer"
 
 import "jest-styled-components"
 
-import AuthorInfo from "../sections/authors"
-import { Authors } from "./fixtures/components"
+import { Authors } from "../../fixtures/components"
+import AuthorInfo from "../authors"
 
 it("renders properly", () => {
   const authors = renderer.create(<AuthorInfo authors={Authors} />).toJSON()
