+++ src/components/publishing/sections/__test__/embed.test.tsx
@@ -3,8 +3,8 @@ import * as renderer from "react-test-renderer"
 
 import "jest-styled-components"
 
-import Embed from "../sections/embed"
-import { Embeds } from "./fixtures/components"
+import { Embeds } from "../../fixtures/components"
+import Embed from "../embed"
 
 it("renders properly", () => {
   const embed = renderer.create(<Embed section={Embeds[0]} />).toJSON()
