+++ src/components/publishing/sections/__test__/imageset_preview.test.tsx
@@ -1,10 +1,9 @@
-import ImageSetPreview from "../sections/imageset_preview"
-import { ImageSetFull, ImageSetMini } from "./fixtures/components"
-
 import "jest-styled-components"
 
 import * as React from "react"
 import * as renderer from "react-test-renderer"
+import { ImageSetFull, ImageSetMini } from "../../fixtures/components"
+import ImageSetPreview from "../imageset_preview"
 
 it("renders a full image set properly", () => {
   const imageset = renderer.create(<ImageSetPreview section={ImageSetFull} />).toJSON()
