+++ src/components/publishing/sections/__test__/imageset_preview_classic.test.tsx
@@ -1,10 +1,9 @@
-import ImageSetPreviewClassic from "../sections/imageset_preview_classic"
-import { Images } from "./fixtures/components"
-
 import "jest-styled-components"
 
 import * as React from "react"
 import * as renderer from "react-test-renderer"
+import { Images } from "../../fixtures/components"
+import ImageSetPreviewClassic from "../imageset_preview_classic"
 
 it("renders properly", () => {
   const imageset = renderer.create(<ImageSetPreviewClassic images={Images} />).toJSON()
