+++ src/components/publishing/sections/__test__/sections.test.tsx
@@ -3,12 +3,12 @@ import * as renderer from "react-test-renderer"
 
 import "jest-styled-components"
 
-import Video from "../sections/video"
-import { Videos } from "./fixtures/components"
+import Articles from "../../fixtures/articles"
+import Sections from "../sections"
 
 jest.mock("react-sizeme", () => jest.fn(c => d => d))
 
 it("renders properly", () => {
-  const video = renderer.create(<Video section={Videos[0]} />).toJSON()
-  expect(video).toMatchSnapshot()
+  const sections = renderer.create(<Sections article={Articles[1]} />).toJSON()
+  expect(sections).toMatchSnapshot()
 })
